Move custom matchers to own file
